move matchers to subdir
